+++ packages/_private/angular-demo/src/AllComponentParts/Select.ts
@@ -67,7 +67,7 @@ import { type SelectItem, mockAsyncData, mockData } from '@synergy-design-system
     </syn-select>
 
     <syn-select
-      data-testid="select-851-multiple"
+      data-testid="select-847-multiple"
       help-text="Normal value binding and async options"
       label="Multiple with async options"
       multiple