+++ packages/_private/vanilla-demo/src/main.ts
@@ -3,6 +3,7 @@
 import {
   type SynChangeEvent,
   type SynIconButton,
+  type SynSideNav,
   type SynSwitch,
   enableExperimentalSettingEmitEvents,
   registerIconLibrary,
@@ -40,6 +41,28 @@ import {
 let currentTheme: AllowedThemes = 'synergy';
 let currentMode: AllowedModes = 'light';
 
+const initLayoutSwitch = async () => {
+  await customElements.whenDefined('syn-button-group');
+
+  const buttons = document.querySelectorAll<SynIconButton>('.meta-navigation syn-button-group syn-icon-button');
+  document
+    .querySelector('.meta-navigation syn-button-group.sidenav-switch')!
+    .addEventListener('click', e => {
+      const target = e.target as HTMLElement;
+      const clickedButton = target.closest('syn-icon-button') as SynIconButton;
+      const type = clickedButton.dataset.type!;
+
+      const sidenav = document.querySelector<SynSideNav>('syn-side-nav');
+      if (sidenav) {
+        sidenav.variant = type as SynSideNav['variant'];
+      }
+
+      buttons.forEach(btn => {
+        btn.color = btn.dataset.type === type ? 'primary' : 'currentColor';
+      });
+    });
+};
+
 const initThemeSwitch = async () => {
   await customElements.whenDefined('syn-icon-button');
   await customElements.whenDefined('syn-switch');
@@ -67,7 +90,7 @@ const initSizeSwitch = async () => {
 
   const buttons = document.querySelectorAll<SynIconButton>('.meta-navigation syn-button-group syn-icon-button');
   document
-    .querySelector('.meta-navigation syn-button-group')!
+    .querySelector('.meta-navigation syn-button-group.size-switch')!
     .addEventListener('click', e => {
       const target = e.target as HTMLElement;
       const clickedButton = target.closest('syn-icon-button') as SynIconButton;
@@ -95,6 +118,7 @@ const bootstrap = async () => {
   document.querySelector('#root')!.innerHTML = createLayout();
 
   await initRouting();
+  await initLayoutSwitch();
   await initThemeSwitch();
   await initSizeSwitch();
 };