+++ packages/components/scripts/vendorism/custom/select.vendorism.js
@@ -251,7 +251,7 @@ const transformComponent = (path, originalContent) => {
     ],
   ], content);
 
-  // #851: Fix angular problem with multiple and initial attribute
+  // #847: Fix angular problem with multiple and initial attribute
   // value ( `<syn-select multiple value="1 2">` )
   content = addSectionsBefore([
     [
@@ -277,7 +277,7 @@ const transformComponent = (path, originalContent) => {
     const value = Array.isArray(val) ? val :  typeof val === 'string' ? val.split(this.delimiter) : [val].filter(Boolean);`,
     ],
   ], content);
-  // End#851
+  // End#847
 
   return {
     content,